Keep relative position between equal screens
When moving between two screens that are of equal size, keep the
relative position the same.

If it's on the left, keep it on the left.

Another way to tackle this would be to *always* keep the relative
positioning between screens, so left half would map to left half, even
moving to screens of different size, but that proportionality is out of
scope for what I need right now.